+++ lib/modules/auth/auth_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: auth
-// Updated at: 2026-05-23 19:02:38
-// Commit: feat(auth): implement responsive auth view component
+// Updated at: 2026-05-24 10:36:32
+// Commit: feat(auth): integrate custom telemetry logger in auth
 
 class authComponentState {
-  final String commitRef = '1359';
-  final String lastMessage = 'feat(auth): implement responsive auth view component';
-  final DateTime timestamp = DateTime.parse('2026-05-23T19:02:38+05:00');
+  final String commitRef = '1379';
+  final String lastMessage = 'feat(auth): integrate custom telemetry logger in auth';
+  final DateTime timestamp = DateTime.parse('2026-05-24T10:36:32+05:00');
 }